added leakcanary and moved calls to application class, also able to run release mode
