+++ sentry-sample/src/main/java/io/sentry/sample/MainActivity.java
@@ -1,7 +1,6 @@
 package io.sentry.sample;
 
 import android.os.Bundle;
-import android.os.StrictMode;
 import androidx.appcompat.app.AppCompatActivity;
 import io.sentry.core.Breadcrumb;
 import io.sentry.core.Sentry;
@@ -14,13 +13,8 @@ public class MainActivity extends AppCompatActivity {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
-    // ideally Application class
-    districtMode();
-
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_main);
-
-    Timber.plant(new Timber.DebugTree());
 
     Timber.i("Sentry.isEnabled() = %s", Sentry.isEnabled());
 
@@ -59,19 +53,7 @@ public class MainActivity extends AppCompatActivity {
                     scope.setUser(user);
                     scope.setTag("tag", "tag");
                   });
+              Sentry.captureException(new Exception("Some exception with scope."));
             });
-  }
-
-  private void districtMode() {
-    //    https://developer.android.com/reference/android/os/StrictMode
-    //    StrictMode is a developer tool which detects things you might be doing by accident and
-    // brings them to your attention so you can fix them.
-
-    if (BuildConfig.DEBUG) {
-      StrictMode.setThreadPolicy(
-          new StrictMode.ThreadPolicy.Builder().detectAll().penaltyLog().build());
-
-      StrictMode.setVmPolicy(new StrictMode.VmPolicy.Builder().detectAll().penaltyLog().build());
-    }
   }
 }
